Create Loding layout
